fixes for LLP win64

+++ src/bltDataTable.c
@@ -5747,7 +5747,7 @@ blt_table_set_row_map(Table *tablePtr, Row **map)
 
 void
 blt_table_sort_init(Table *tablePtr, BLT_TABLE_SORT_ORDER *order, 
-                    size_t numColumns, unsigned int flags)
+                    long numColumns, unsigned int flags)
 {
     size_t i;
 
@@ -5779,7 +5779,7 @@ blt_table_sort_rows(Table *tablePtr)
 
 
 void
-blt_table_sort_row_map(Table *tablePtr, size_t numRows, BLT_TABLE_ROW *rows)
+blt_table_sort_row_map(Table *tablePtr, long numRows, BLT_TABLE_ROW *rows)
 {
     /* Sort the map and return it. */
     qsort((char *)rows, numRows, sizeof(BLT_TABLE_ROW),
